requirements.txt
langchain
langgraph
langchain-core
langchain-community
ipykernel 
python-dotenv
langchain-groq
Ipython
arxiv
wikipedia
langchain-together
streamlit